// saas-restaurants/src/routers/type.ts
export enum RouteLinks {
    LOGIN = '/login',
    DASHBOARD = '/',
    RESTAURANTS = '/restaurants',
    TENANT = '/tenant',
    SUBSCRIPTIONS = '/subscriptions',
    INVOICES = '/invoices',
    USERS = '/users',
    REPORTS = '/reports',
    NOTIFICATIONS = '/notifications',
    SETTINGS = '/settings',
    UNAUTHORIZED = '/unauthorized',
    RESTAURANT_DETAIL = '/restaurants/:tenantId/:id/:type'
}